Added sweetalert2 for jobs/resumes

+++ client/src/app/RESUMES/resume/resume.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ResumeService } from '../resume.service';
-// import { ToastrService } from 'ngx-toastr';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-resume',
@@ -27,18 +27,27 @@ export class ResumeComponent implements OnInit {
   }
 
   deleteData(id) {
-    this.resumeService.deleteData(id).subscribe((res) => {
-      this.data = res;
-      // this.toastr.error(
-      //   JSON.stringify(this.data.code),
-      //   JSON.stringify(this.data.message),
-      //   {
-      //     timeOut: 2000,
-      //     progressBar: true,
-      //   }
-      // );
-      alert('Resume Deleted');
-      this.getResumesData();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this resume!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.resumeService.deleteData(id).subscribe((res) => {
+          this.data = res;
+
+          this.getResumesData();
+        });
+
+        Swal.fire('Deleted!', 'Your resume has been deleted.', 'success');
+        // For more information about handling dismissals please visit
+        // https://sweetalert2.github.io/#handling-dismissals
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire('Cancelled', 'Your resume is safe :)', 'error');
+      }
     });
   }
 }
